# full-stack/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VaultPay API
  version: 1.0.0
  description: MVP API for a partner-led Nigerian financial platform. Configure BANK_PROVIDER and KYC_PROVIDER to use the Flutterwave adapter; cards remain issuer/provider dependent.
servers:
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/register:
    post:
      security: []
      responses: {'201': {description: User created}, '409': {description: User exists}}
...